+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 - It’s now possible to import Single sections’ entry data. ([#559](https://github.com/craftcms/feed-me/issues/559))
 - It’s now possible to import global set data. ([#670](https://github.com/craftcms/feed-me/issues/670))
+- Added the `feed-me/feeds/queue` command, which accepts a comma-separated list of feed IDs, and queues them up to be processed. ([#754](https://github.com/craftcms/feed-me/issues/754))
 - Added `craft\feedme\elements\GlobalSet`.
 - Added `craft\feedme\models\ElementGroup`.
 - Added `craft\feedme\models\FeedModel::$singleton`.
@@ -12,10 +13,12 @@
 ### Changed
 - Feed Me now requires Craft 3.5 or later.
 - Enforce stricter validation when setting up a feed to help prevent unwanted behaviors. ([#757](https://github.com/craftcms/feed-me/issues/757))
-- The `php craft feed-me/feeds/run` console command now takes a list of feed IDs to run as a parameter instead of using an `--id` flag. (i.e. `php craft feed-me/feeds/run 1,2`). ([#754](https://github.com/craftcms/feed-me/issues/754))
 - Updated `ralouphie/mimey` to 2.1.0. ([#740](https://github.com/craftcms/feed-me/issues/740))
 - Cleaned up log messages when there are no items in a feed to process. ([#585](https://github.com/craftcms/feed-me/issues/585))
 - Elements’ `getGroups()` methods can now return `craft\feedme\models\ElementGroup` objects.
+
+### Removed
+- Removed the `feed-me/feeds/run` command. The new `feed-me/feeds/queue` command can be used instead, in combination with `queue/run`.
 
 ### Fixed
 - Fixed a SQL error that would occur when trying to import Products in Commerce 3.2.8 or higher.